Create Compositional Layout extensions
