+++ spotty/spotty/UIComponents/Layout/NSCollectionLayoutBoundarySupplementaryItem+createSectionHeader.swift
@@ -1,0 +1,23 @@
+//
+//  NSCollectionLayoutBoundarySupplementaryItem+createSectionHeader.swift
+//  spotty
+//
+//  Created by Ivaylo Kalaydzhiev on 20.06.22.
+//
+
+import UIKit
+
+extension NSCollectionLayoutBoundarySupplementaryItem {
+    
+    static func createSectionHeader() -> NSCollectionLayoutBoundarySupplementaryItem {
+        let layoutSectionHeaderSize = NSCollectionLayoutSize(
+            widthDimension: .fractionalWidth(0.93),
+            heightDimension: .estimated(80))
+        let layoutSectionHeader = NSCollectionLayoutBoundarySupplementaryItem(
+            layoutSize: layoutSectionHeaderSize,
+            elementKind: UICollectionView.elementKindSectionHeader,
+            alignment: .top)
+
+        return layoutSectionHeader
+    }
+}
